requirements.txt
Django==4.0.4
gunicorn==20.1.0
pypng==0.0.21
PyQRCode==1.2.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.1.0
